refactor: util 오타 수정
