+++ src/main/java/racingcar/util/Validator.java
@@ -1,4 +1,4 @@
-package racingcar.utill;
+package racingcar.util;
 
 import java.util.Arrays;
 import java.util.List;
