+++ src/main/java/racingcar/view/InputView.java
@@ -3,7 +3,7 @@ package racingcar.view;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
 import java.util.List;
-import racingcar.utill.Validator;
+import racingcar.util.Validator;
 
 public class InputView {
